app/requirements.txt
Django==5.2.6
psycopg2-binary==2.9.9
gunicorn==23.0.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
Pillow==11.2.1
deepface==0.0.93
tf-keras==2.19.0
reportlab==4.4.1
